Add files via upload
this is with out full set with bbox
we can see 3 set 0-1-2
we can see at each set 4 read new line cause mem has 4 lines to read. but we would like it to be 1'b1 only 3 times , cause start read is the first read request. but the mem buffer will stop at the end of line even thougth we have new request: read new line. (he will read but we won't use this). and the last set isn't full.