Add files via upload
for 70 bbox:
in the first 2 set, we have 24 bbox in each one, this is 2 state of select pe state.
in the last set we remain with 22 bbox, therefore we need 5 state of pe 0 and one state of select pe 2

for 24 bbox:
there is one full set, so we need only select pe state. pe sel :0-5, row_sel=0

for 16 bbox:
there is one set, that is not full, therefore we need to move directly to pe0state and we need him 4 times cause: 16/4=4: pe_sel: 0-3, row sel=0


for 15 bbox:
there is one set, that is not full, therefore we need to move directly to pe0state and we need him 3 times cause: 15/4=3: pe_sel: 0-2, row sel=0, and we need pe3 state cause 15%4=3.


for 4 bbox:
there is one set, that is not full, therefore we need to move directly to pe0state and we need him 1 times cause: 4/4=1: pe_sel: 0, row sel=0.


for 3 bbox:
there is one set, that is not full, and there is less than 4 bbox,therefore we need to move directly to pe3state and we need him 1 times cause: pe_sel: 0, row sel=0, remainder=3.